refactor: move audiocall & sprint contexts to separate folder
